Add check valid host test for new streams
Adds a function named check_valid_host to run some tests on the
host to ensure that a stream can successfully open using it. This
also attempts to solve the issue where local testing using multiple
nodes on cluster_run sends back $HOST inside the config, which
cannot be streamed on.